refactor: navigate to company page by id when click to button on dashboard
