+++ src/components/Dashboard/PopularCompanies/PopularCompanies.tsx
@@ -10,7 +10,8 @@ import ListItemAvatar from '@mui/material/ListItemAvatar';
 import ListItemText from '@mui/material/ListItemText';
 import type { SxProps } from '@mui/material/styles';
 import ArrowForwardIcon from '@mui/icons-material/ArrowForward';
-import PopularCompaniesModal from '../PopularCompaniesModal/PopularCompaniesModal.tsx';
+import PopularCompaniesModal from '../../PopularCompaniesModal/PopularCompaniesModal.tsx';
+import { useNavigate } from 'react-router-dom';
 
 export interface Company {
   id: string;
@@ -25,6 +26,11 @@ export interface LatestProductsProps {
 
 export function PopularCompanies({ companies = [], sx }: LatestProductsProps) {
   const [open, setOpen] = useState(false);
+  const navigate = useNavigate();
+
+  const goToCompanyPage = (companyId: string) => {
+    navigate(`/companypage/${companyId}`); // Navigate to the company's page
+  };
 
   return (
     <Card sx={sx}>
@@ -59,7 +65,7 @@ export function PopularCompanies({ companies = [], sx }: LatestProductsProps) {
                 secondaryTypographyProps={{ variant: 'body2' }}
               />
 
-              <Button variant="contained" endIcon={<ArrowForwardIcon />}>
+              <Button variant="contained" onClick={() => goToCompanyPage(company.id)} endIcon={<ArrowForwardIcon />}>
                 explore
               </Button>
             </ListItem>
